Started programming the renderer and vulkan init

+++ engine/src/game_interface.h
@@ -5,17 +5,18 @@
 #pragma once
 
 #include "core/application.h"
+#include "renderer/renderer_backend.h"
 
 typedef struct Game {
   ApplicationInfo application_info;
 
   b8 (*initialize)(struct Game* game_instance);
 
-  b8 (*on_update)(struct Game* game_instance, f32 delta_time);
+  b8 (*on_update)(struct Game* game_instance, f64 delta_time);
 
-  b8 (*on_render)(struct Game* game_instance, f32 delta_time);
+  b8 (*on_render)(struct Game* game_instance, f64 delta_time);
 
-  void (*on_resize)(struct Game* game_instance, u32 width, u32 height);
+  void (*on_resize)(struct Game* game_instance, u16 width, u16 height);
 
   void* state;
