+++ test/src/game.h
@@ -8,7 +8,7 @@
 #include <game_interface.h>
 
 typedef struct GameState {
-  f32 delta_time;
+  f64 delta_time;
 } GameState;
 
 /**
@@ -24,8 +24,8 @@ b8 game_initalize(Game* game_instance);
  * @param delta_time The time that the update is called.
  * @brief Whether the update was succesful.
  */
-b8 on_update(Game* game_instance, f32 delta_time);
+b8 on_update(Game* game_instance, f64 delta_time);
 
-b8 on_render(Game* game_instance, f32 delta_time);
+b8 on_render(Game* game_instance, f64 delta_time);
 
-void on_resize(Game* game_instance, u32 width, u32 height);
+void on_resize(Game* game_instance, u16 width, u16 height);
